Refactor sync committee helpers
