+++ tests/core/pyspec/eth2spec/test/helpers/sync_committee.py
